Made a web assembly version. Still having problems doing that with Pixel Game Engine so I eventually made one in SDL2.

+++ olc_code_jam_2021/src/jam/TileMap.h
@@ -21,7 +21,7 @@ namespace jam
 		int GetWidth() { return this->width * this->tileWidth; };
 		int GetHeight() { return this->height * this->tileHeight; };
 		TileMapLayer* GetLayer(int layerNum) { return this->layers[layerNum]; };
-		int GetLayerCount() { return this->layers.size(); };
+		int GetLayerCount() { return (int)this->layers.size(); };
 		int GetTileWidth() { return this->tileWidth; };
 		int GetTileHeight() { return this->tileHeight; };
 	private:
